requirements.txt
Flask==1.1.2
Jinja2==2.11.2
mysql.connector==2.2.9